refactor(web): relocate assets directory
Moved all asset files from the `src/assets` directory to the `assets` directory under the `apps/web` module. Updated import paths across affected components to reflect the new structure. This change improves consistency and better segregates assets from source files.
