+++ apps/web/assets/hero/giovanni_ravalico-profile_color_e4cily_c_scale,w_929.webp
